requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-auth==2.29.0
requests==2.31.0
qrcode==7.4.2
Pillow==10.3.0
SQLAlchemy==2.0.30
python-dateutil==2.9.0
gunicorn==22.0.0